refactor(server): change server structure
